# workspaces/P20260914-1728-网页版票夹管理发票/03_engineering/scaffold/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponseAttachmentView": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AttachmentView"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListAttachmentView": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AttachmentView"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "AttachmentView": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/attachments/{id}": {
      "delete": {
        "operationId": "deleteAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除单个附件（R7：不影响其他）",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/v1/attachments/{id}/file": {
      "get": {
        "operationId": "downloadAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "下载/预览（属主谓词）",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/api/v1/invoices/{id}/attachments": {
      "get": {
        "operationId": "listAttachments",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAttachmentView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "附件列表",
        "tags": [
          "attachments"
        ]
      },
      "post": {
        "description": "三校验：>10MB→ATT_001；已挂3个→ATT_002；格式/魔数不符→ATT_003",
        "operationId": "uploadAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAttachmentView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "上传附件（F11）",
        "tags": [
          "attachments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}